perf(distributions): return array instead of object (#209)
Makes `bernoulliDistribution`, `binomialDistribution` and
`poissonDistribution` return an array instead of object. The result is
fully equivalent since keys were sequential integers starting from 0,
but this is ~5% faster.

BREAKING CHANGE: The return value of bernoulliDistribution, binomialDistribution, and poissonDistribution is no longer an Object with number keys, it is now an Array of numbers.