docs(readme): Close script tag
In HTML5, script is not a void element [1] and therefore needs to be
closed, otherwise it "eats" whatever comes after it (until </script>
is encountered).

This might not apply to XML-serialized HTML5 [2], but I don't know
anyone who uses that.

Note that even for void elements, trailing slashes are not recommended.
[3] Presumably because they are optional, and because they lead people
to believe that normal elements can also be self-closed.

The closing tag is on a new line, because otherwise GitHub adds an ugly
horizontal scroll bar to the code block.

[1] https://www.w3.org/TR/html5/syntax.html#void-elements
[2] https://mathiasbynens.be/notes/xhtml5
[3] http://codeguide.co/#html-syntax